requirements.txt
 requests
 pandas
 ta
 python-telegram-bot